+++ example/searchActions.js
@@ -1,15 +1,13 @@
 /**
  * アクション検索サンプル
- * @ignore
  */
-
 const moment = require('moment');
 const pecorino = require('../');
 
-pecorino.mongoose.connect(process.env.MONGOLAB_URI);
+mongoose.connect(process.env.MONGOLAB_URI);
 
 async function main() {
-    const actionRepo = new pecorino.repository.Action(pecorino.mongoose.connection);
+    const actionRepo = new pecorino.repository.Action(mongoose.connection);
     const actions = await actionRepo.search({
         typeOf: pecorino.factory.actionType.MoneyTransfer,
         startDateFrom: moment().add(-1, 'hour').toDate(),
@@ -27,5 +25,5 @@ main().then(() => {
     console.log('success!');
 }).catch(console.error)
     .then(() => {
-        pecorino.mongoose.disconnect();
+        mongoose.disconnect();
     });