+++ src/service/transaction/deposit.spec.ts
@@ -1,8 +1,8 @@
 // tslint:disable:no-implicit-dependencies
 /**
  * 入金取引サービステスト
- * @ignore
  */
+import * as mongoose from 'mongoose';
 import * as assert from 'power-assert';
 import * as sinon from 'sinon';
 
@@ -25,13 +25,16 @@ describe('入金取引を開始する', () => {
             typeOf: pecorino.factory.transactionType.Deposit,
             agent: {},
             recipient: {},
-            object: {},
+            object: {
+                fromLocation: {},
+                toLocation: {}
+            },
             expires: new Date(),
             status: pecorino.factory.transactionStatusType.Confirmed,
             startDate: new Date()
         };
-        const accountRepo = new pecorino.repository.Account(pecorino.mongoose.connection);
-        const transactionRepo = new pecorino.repository.Transaction(pecorino.mongoose.connection);
+        const accountRepo = new pecorino.repository.Account(mongoose.connection);
+        const transactionRepo = new pecorino.repository.Transaction(mongoose.connection);
         sandbox.mock(accountRepo).expects('findByAccountNumber').once().resolves(account);
         sandbox.mock(transactionRepo).expects('start').once().resolves(transaction);
         sandbox.mock(accountRepo).expects('startTransaction').once().resolves();
@@ -53,14 +56,17 @@ describe('入金取引を開始する', () => {
             typeOf: pecorino.factory.transactionType.Deposit,
             agent: {},
             recipient: {},
-            object: {},
+            object: {
+                fromLocation: {},
+                toLocation: {}
+            },
             expires: new Date(),
             status: pecorino.factory.transactionStatusType.Confirmed,
             startDate: new Date()
         };
         const startError = new Error('startError');
-        const accountRepo = new pecorino.repository.Account(pecorino.mongoose.connection);
-        const transactionRepo = new pecorino.repository.Transaction(pecorino.mongoose.connection);
+        const accountRepo = new pecorino.repository.Account(mongoose.connection);
+        const transactionRepo = new pecorino.repository.Transaction(mongoose.connection);
         sandbox.mock(accountRepo).expects('findByAccountNumber').once().resolves(account);
         sandbox.mock(transactionRepo).expects('start').once().rejects(startError);
         sandbox.mock(accountRepo).expects('startTransaction').never();
@@ -88,13 +94,16 @@ describe('入金取引を確定する', () => {
             typeOf: pecorino.factory.transactionType.Deposit,
             agent: {},
             recipient: {},
-            object: {},
+            object: {
+                fromLocation: {},
+                toLocation: {}
+            },
             expires: new Date(),
             status: pecorino.factory.transactionStatusType.Confirmed,
             result: {},
             startDate: new Date()
         };
-        const transactionRepo = new pecorino.repository.Transaction(pecorino.mongoose.connection);
+        const transactionRepo = new pecorino.repository.Transaction(mongoose.connection);
         sandbox.mock(transactionRepo).expects('findById').once().resolves(transaction);
         sandbox.mock(transactionRepo).expects('confirm').once().resolves(transaction);
 
@@ -119,8 +128,8 @@ describe('取引のタスクをエクスポートする', () => {
             status: pecorino.factory.transactionStatusType.Canceled
         };
         const task = {};
-        const taskRepo = new pecorino.repository.Task(pecorino.mongoose.connection);
-        const transactionRepo = new pecorino.repository.Transaction(pecorino.mongoose.connection);
+        const taskRepo = new pecorino.repository.Task(mongoose.connection);
+        const transactionRepo = new pecorino.repository.Transaction(mongoose.connection);
         sandbox.mock(transactionRepo).expects('startExportTasks').once().resolves(transaction);
         sandbox.mock(transactionRepo).expects('findById').once().resolves(transaction);
         sandbox.mock(taskRepo).expects('save').atLeast(1).resolves(task);
@@ -135,8 +144,8 @@ describe('取引のタスクをエクスポートする', () => {
     });
 
     it('タスクエクスポート待ちの取引がなければ何もしないはず', async () => {
-        const taskRepo = new pecorino.repository.Task(pecorino.mongoose.connection);
-        const transactionRepo = new pecorino.repository.Transaction(pecorino.mongoose.connection);
+        const taskRepo = new pecorino.repository.Task(mongoose.connection);
+        const transactionRepo = new pecorino.repository.Transaction(mongoose.connection);
         sandbox.mock(transactionRepo).expects('startExportTasks').once().resolves(null);
         sandbox.mock(transactionRepo).expects('findById').never();
         sandbox.mock(taskRepo).expects('save').never();
@@ -165,8 +174,8 @@ describe('ID指定で取引のタスクをエクスポートする', () => {
             }
         };
         const task = {};
-        const taskRepo = new pecorino.repository.Task(pecorino.mongoose.connection);
-        const transactionRepo = new pecorino.repository.Transaction(pecorino.mongoose.connection);
+        const taskRepo = new pecorino.repository.Task(mongoose.connection);
+        const transactionRepo = new pecorino.repository.Transaction(mongoose.connection);
         sandbox.mock(transactionRepo).expects('findById').once().resolves(transaction);
         sandbox.mock(taskRepo).expects('save').atLeast(1).resolves(task);
 
@@ -189,8 +198,8 @@ describe('ID指定で取引のタスクをエクスポートする', () => {
                 status: transactionStatus
             };
             const task = {};
-            const taskRepo = new pecorino.repository.Task(pecorino.mongoose.connection);
-            const transactionRepo = new pecorino.repository.Transaction(pecorino.mongoose.connection);
+            const taskRepo = new pecorino.repository.Task(mongoose.connection);
+            const transactionRepo = new pecorino.repository.Transaction(mongoose.connection);
             sandbox.mock(transactionRepo).expects('findById').once().resolves(transaction);
             sandbox.mock(taskRepo).expects('save').atLeast(1).resolves(task);
 
@@ -208,8 +217,8 @@ describe('ID指定で取引のタスクをエクスポートする', () => {
             id: 'transactionId',
             status: 'UnknownStatus'
         };
-        const taskRepo = new pecorino.repository.Task(pecorino.mongoose.connection);
-        const transactionRepo = new pecorino.repository.Transaction(pecorino.mongoose.connection);
+        const taskRepo = new pecorino.repository.Task(mongoose.connection);
+        const transactionRepo = new pecorino.repository.Transaction(mongoose.connection);
         sandbox.mock(transactionRepo).expects('findById').once().resolves(transaction);
         sandbox.mock(taskRepo).expects('save').never();