+++ src/service/transaction/transfer.ts
@@ -36,12 +36,12 @@ export function start<T extends factory.account.AccountType>(
 
         // 口座存在確認
         const fromAccount = await repos.account.findByAccountNumber<T>({
-            accountType: params.object.accountType,
-            accountNumber: params.object.fromAccountNumber
+            accountType: params.object.fromLocation.accountType,
+            accountNumber: params.object.fromLocation.accountNumber
         });
         const toAccount = await repos.account.findByAccountNumber<T>({
-            accountType: params.object.accountType,
-            accountNumber: params.object.toAccountNumber
+            accountType: params.object.toLocation.accountType,
+            accountNumber: params.object.toLocation.accountNumber
         });
 
         // 取引ファクトリーで新しい進行中取引オブジェクトを作成
@@ -52,10 +52,19 @@ export function start<T extends factory.account.AccountType>(
             object: {
                 clientUser: params.object.clientUser,
                 amount: params.object.amount,
-                accountType: params.object.accountType,
-                fromAccountNumber: fromAccount.accountNumber,
-                toAccountNumber: toAccount.accountNumber,
-                notes: params.object.notes
+                fromLocation: {
+                    typeOf: factory.account.TypeOf.Account,
+                    accountType: fromAccount.accountType,
+                    accountNumber: fromAccount.accountNumber,
+                    name: fromAccount.name
+                },
+                toLocation: {
+                    typeOf: factory.account.TypeOf.Account,
+                    accountType: toAccount.accountType,
+                    accountNumber: toAccount.accountNumber,
+                    name: toAccount.name
+                },
+                description: params.object.description
             },
             expires: params.expires
         };
@@ -82,16 +91,16 @@ export function start<T extends factory.account.AccountType>(
 
         // 残高確認
         await repos.account.authorizeAmount<T>({
-            accountType: params.object.accountType,
-            accountNumber: params.object.fromAccountNumber,
+            accountType: params.object.fromLocation.accountType,
+            accountNumber: params.object.fromLocation.accountNumber,
             amount: params.object.amount,
             transaction: pendingTransaction
         });
 
         // 転送先口座に進行中取引を追加
         await repos.account.startTransaction<T>({
-            accountType: params.object.accountType,
-            accountNumber: params.object.toAccountNumber,
+            accountType: params.object.toLocation.accountType,
+            accountNumber: params.object.toLocation.accountNumber,
             transaction: pendingTransaction
         });
 
@@ -117,7 +126,7 @@ export function confirm<T extends factory.account.AccountType>(params: {
         // 現金転送アクション属性作成
         const moneyTransferActionAttributes: factory.action.transfer.moneyTransfer.IAttributes<T> = {
             typeOf: factory.actionType.MoneyTransfer,
-            description: transaction.object.notes,
+            description: transaction.object.description,
             result: {
                 amount: transaction.object.amount
             },
@@ -127,15 +136,11 @@ export function confirm<T extends factory.account.AccountType>(params: {
             recipient: transaction.recipient,
             amount: transaction.object.amount,
             fromLocation: {
-                typeOf: factory.account.TypeOf.Account,
-                accountType: transaction.object.accountType,
-                accountNumber: transaction.object.fromAccountNumber,
+                ...transaction.object.fromLocation,
                 name: transaction.agent.name
             },
             toLocation: {
-                typeOf: factory.account.TypeOf.Account,
-                accountType: transaction.object.accountType,
-                accountNumber: transaction.object.toAccountNumber,
+                ...transaction.object.toLocation,
                 name: transaction.recipient.name
             },
             purpose: {
@@ -199,8 +204,6 @@ export function exportTasksById<T extends factory.account.AccountType>(
                             status: factory.taskStatus.Ready,
                             runsAt: new Date(), // なるはやで実行
                             remainingNumberOfTries: 10,
-                            // tslint:disable-next-line:no-null-keyword
-                            lastTriedAt: null,
                             numberOfTried: 0,
                             executionResults: [],
                             data: {
@@ -219,8 +222,6 @@ export function exportTasksById<T extends factory.account.AccountType>(
                     status: factory.taskStatus.Ready,
                     runsAt: new Date(), // なるはやで実行
                     remainingNumberOfTries: 10,
-                    // tslint:disable-next-line:no-null-keyword
-                    lastTriedAt: null,
                     numberOfTried: 0,
                     executionResults: [],
                     data: {